backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
httpx==0.28.0
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
maxminddb==2.6.2
geoip2==4.8.0
user-agents==2.2.0
slowapi==0.1.9
bleach==6.2.0
aiosmtplib==3.0.1
